added godfather

+++ engine/roles/godfather.py
@@ -1,5 +1,6 @@
 
 from typing import List
+import random
 
 import engine.events as events
 from engine.events import ACTION_EVENTS
@@ -7,10 +8,10 @@ import engine.roles as roles
 from engine.utils.logger import logger
 
 
-class Mafioso(roles.Actor):
+class Godfather(roles.Actor):
     def __init__(self, player: dict, settings: dict):
         super().__init__(player)
-        self.role_name = 'Mafioso'
+        self.role_name = 'Godfather'
         self.alignment = roles.Alignment.MAFIA
         self.kill_message = "You were killed by a member of the Mafia" # TODO
         self.death_reason = "They were found riddled with bullets" # TODO
@@ -19,28 +20,27 @@ class Mafioso(roles.Actor):
         self.allies = [actor for actor in actors if actor.alignment == self.alignment]
         return
     
-    # def find_possible_targets(self, actors: List[roles.Actor] = None) -> None:
-    #     # Number of targets
-    #     num_targets = 1 
-    #     self.possible_targets = []
-    #     for i in range(num_targets):
-    #         self.possible_targets.insert(i, [
-    #             actor for actor in actors 
-    #             if actor.alive
-    #             and actor.alignment != self.alignment
-    #             and actor.number != self.number # Seems a bit redundant, but can't hurt
-    #         ])
+    def find_possible_targets(self, actors: List[roles.Actor] = None) -> None:
+        # Number of targets
+        num_targets = 1 
+        self.possible_targets = []
+        for i in range(num_targets):
+            self.possible_targets.insert(i, [
+                actor for actor in actors 
+                if actor.alive
+                and actor.alignment != self.alignment
+                and actor.number != self.number # Seems a bit redundant, but can't hurt
+            ])
             
     def action(self, targets: List[roles.Actor]=[]):
         target = targets[0]
         def success():
-            print('Target was killed')
-            success_event_group = events.GameEventGroup(group_id='mafioso_action_success', duration=events.Duration.MAFIA_KILL)
+            success_event_group = events.GameEventGroup(group_id='godfather_action_success', duration=events.Duration.MAFIA_KILL)
             
             # Inform all players that a Mafia kill has succeeded
             success_event_group.new_event(
                 events.GameEvent(
-                    event_id="mafia_kill_success",
+                    event_id="godfather_kill_success",
                     targets=['*'],
                     message="There are sounds of shots in the streets"
                 )
@@ -59,12 +59,12 @@ class Mafioso(roles.Actor):
         
         def fail():
             print('Target survived')
-            fail_event_group = events.GameEventGroup(group_id="mafioso_action_fail", duration=events.Duration.MAFIA_KILL)
+            fail_event_group = events.GameEventGroup(group_id="godfather_action_fail", duration=events.Duration.MAFIA_KILL)
             
             # Inform all players that a Mafia kill has failed
             fail_event_group.new_event(
                 events.GameEvent(
-                    event_id='mafia_kill_fail',
+                    event_id='godfather_kill_fail',
                     targets=['*'],
                     message=''
                 )
@@ -72,5 +72,23 @@ class Mafioso(roles.Actor):
             
             ACTION_EVENTS.new_event_group(fail_event_group)
         
-        
-        self.kill(target, success, fail)
+        # Check if there are idle Mafioso that you can send, else go yourself
+        proxies = [ally for ally in self.allies if ally.role_name == 'Mafioso']
+        if not proxies:
+            self.kill(target, success, fail)
+        else:
+            proxy = random.choice(proxies)
+            # TODO: If not target.witched
+            proxy.targets = self.targets
+            
+            proxy_event_group = events.GameEventGroup(group_id='godfather_proxy')
+            proxy_event_group.new_event(
+                events.GameEvent(
+                    event_id='godfather_proxy_choice',
+                    targets=[ally.player['id'] for ally in self.allies],
+                    message=f'The Godfather has chosed {proxy.alias} to carry out the hit'
+                )
+            )
+            
+            logger.info(f"{self} has chosen {proxy} to act as a proxy")
+            ACTION_EVENTS.new_event_group(proxy_event_group)
